delete refactor using authentication middleware

+++ src/domain/use-cases/product/delete-product-usecase.ts
@@ -7,7 +7,18 @@ export class DeleteProductUseCase {
     this.productRepository = productRepository;
   }
 
-  async execute(productId: string): Promise<void> {
+  async execute(productId: string, userId: string): Promise<void> {
+    // es el usuario que hace la petición el mismo que ha creado el producto?
+
+    // necesitamos el product que tenemos que borrar
+    const productToRemove = await this.productRepository.findById(productId);
+
+    if (!productToRemove) {
+      throw new Error('Product not found');
+    }
+    if (userId !== productToRemove.ownerId) {
+      throw new Error('Forbidden operation');
+    }
     const isRemoved = await this.productRepository.removeById(productId);
 
     if (!isRemoved) {
